Add files via upload
This is an interactive spray chart that I created based off of a teammates Rapsodo CSV file. This allows the hitter to view all his batted balls on our home field with the same dimensions. The dots are color coded based on if they were hard-hit(green=yes/red=no), and the two gold dots indicate balls that were hard-hit, with proper backspin and spin rate.